refactor and add collapse
